separate the client response

+++ user-service/src/main/java/com/userservice/entity/User.java
@@ -14,8 +14,8 @@ import lombok.NoArgsConstructor;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
     private String name;
     private String email;
-    private Integer numberOfOrders;
+    private int numberOfOrders;
 }
